Refactor paper trading reporter logging and add unit test
